+++ models/lanepartner.js
@@ -10,13 +10,27 @@ module.exports = (sequelize, DataTypes) => {
      * The `models/index` file will call this method automatically.
      */
     static associate(models) {
-      LanePartner.hasOne(models.LanePartnerLocation, {
+      LanePartner.hasOne(models.CustomerLane, {
         foreignKey: 'lanePartnerId'
       })
     }
   };
   LanePartner.init({
-    name: DataTypes.STRING
+    name: DataTypes.STRING,
+    address: DataTypes.STRING,
+    address2: DataTypes.STRING,
+    city: DataTypes.STRING,
+    state: DataTypes.STRING,
+    zipcode: DataTypes.STRING,
+    lnglat: DataTypes.STRING,
+    open: DataTypes.STRING,
+    close: DataTypes.STRING,
+    firstName: DataTypes.STRING,
+    lastName: DataTypes.STRING,
+    title: DataTypes.STRING,
+    phone: DataTypes.STRING,
+    phoneExt: DataTypes.STRING,
+    email: DataTypes.STRING
   }, {
     sequelize,
     modelName: 'LanePartner',
